+++ ChangeLog.md
@@ -1,5 +1,8 @@
 ## ChangeLog - Most recent changes at the top
 
+Version 1.604 - Localization of main .conf and require files
+- MANDATORY UPDATE: hudclass.lua, apclass.lua, customapclass.lua and customhudclass.lua (If using modular)
+
 Version 1.603 - Override Function support for require files
 (Default hudclass.lua and apclass.lua updated)
 - FEATURE: Provided support for Override require files to override specific functions without changing the default require file. (Thanks Davemane42 for format)